+++ lib/src/_internal/transport/polling_transport.dart
@@ -159,6 +159,9 @@ class PollingTransport implements RealtimeTransport {
   }) {}
 
   @override
+  void sendDelivered(String roomId, String messageId) {}
+
+  @override
   void sendMessage(
     String roomId, {
     String? text,