+++ lib/src/_internal/transport/sse_transport.dart
@@ -70,6 +70,9 @@ class SseTransport implements RealtimeTransport {
   }) {}
 
   @override
+  void sendDelivered(String roomId, String messageId) {}
+
+  @override
   void sendMessage(
     String roomId, {
     String? text,